backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "An app that tracks Chicago alderman newsletters"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "html2text>=2025.4.15",
    "imap-tools>=1.11.0",
    "ollama>=0.6.1",
    "pydantic>=2.12.5",
    "pyjwt>=2.10.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "resend>=2.19.0",
    "supabase>=2.27.0",
    "tzdata>=2025.3",
]

[dependency-groups]
dev = [
    "coverage>=7.13.1",
    "mypy>=1.13.0",
    "ruff>=0.14.14",
]
